Fixing gcc warnings when -pedantic is used
Fixing warnings related to casting object pointer to function pointer
and back. Apparently `void*` and `void(*)()` are not the same size on
some weird architectures and gcc complains when casting between the two
when using -pedantic flag.

Work around is to go via uintptr_t. I added helper methods to `cast.h`
and used them in `pybind11.h` and in `numpy.h`. I only tested compilation
with gcc on Linux, but it should be portable to any c++11 compiler.