Add support for implicit conversions to C++ types
See issue #259.

This commit adds support for py::implicitly_convertible<From,To>() to
perform implicit conversion, at the C++ level, when To is not a
pybind11-registered type (e.g. a custom C++ type that should not be
exposed, or a primitive type (double, etc.)).

In essence, this lets you make use of C++ implicit converters (typically
via a 'From::operator To()' or a non-explicit 'To::To(const From&)'
constructor) with pybind11 instances without needing to create an
explicit pybind11 interface for the conversion.

As a simple example, consider the two classes:

    class A {
        ...
        void method(uint64_t val) { ... }
    }
    class B {
        ...
        operator uint64_t() { ... } // returns unique identifier
    }

In C++, you can call `a.method(22)` or `a.method(b)`.  Telling
pybind11 about the conversion allows you to do the same in python.

Without the implicit conversion, you would have to either overload
A.method to accept both uint64_t and B instances, or would have to add
a method to B's interface that returns the id, then pass this method
result to A's method.