Make buffer_info movable but not copyable
buffer_info currently gets a default copy constructor and copy
assignment operator, but not a default move constructor or move
assignment operator.  Nevertheless it is not actually copy-safe (a copy
will result in two pointers to `view`, which will result in a
double-free during deletion (as reported in #377)).

I think this was technically broken already: any code returning a
buffer_info constructed from a view was implicitly depending on return
value copy elision, but a non-eliding compilation should result in
double frees.

This commit makes `view` a unique_ptr, and moves the view destruction
code into a custom deleter of that unique_ptr.  These two changes have
two effects:

- we get a default, implicit move constructor and move assignment
  operator (they were not present before because of the user-declared
  destructor).
- we no longer have (broken) default copy constructor/copy assignment
  (because `view`, now a unique_ptr instead of a raw pointer, is no
  longer copyable).