WIP: PyPy support
This commit includes a few minor modifications to pybind11 that are
needed to get simple hello-world style functions to compile and run on
the latest PyPy.

Sadly, more complex things are still broken: for instance, creating new
types fails with the error message 'TypeError: can't set attributes on
type object <..>' when pybind11 tries to set the __module__ attribute.
Digging into the pip codebase indicates that it believes that the
underlying type is not a heap type, which is incorrect. So this is
likely a PyPy bug.